feat: renaming tools to components
