+++ lib/advert/ui/components/advert_card.dart
