+++ lib/advert/ui/pages/main_page/main_page.dart
@@ -9,8 +9,8 @@ import 'package:myecl/advert/providers/announcer_provider.dart';
 import 'package:myecl/advert/providers/is_advert_admin_provider.dart';
 import 'package:myecl/advert/ui/pages/advert.dart';
 import 'package:myecl/advert/router.dart';
-import 'package:myecl/advert/ui/tools/announcer_bar.dart';
-import 'package:myecl/advert/ui/tools/advert_card.dart';
+import 'package:myecl/advert/ui/components/announcer_bar.dart';
+import 'package:myecl/advert/ui/components/advert_card.dart';
 import 'package:myecl/tools/ui/refresher.dart';
 import 'package:qlevar_router/qlevar_router.dart';
 import 'package:myecl/advert/tools/constants.dart';
